% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{checknames}
\alias{checknames}
\title{check for valid model terms if 'expnms' parameter not given}
\usage{
checknames(terms)
}
\arguments{
\item{terms}{model terms from attr(terms(modelfunction, data), "term.labels")}
}
\description{
This is an internal function called by \code{\link[qgcomp]{qgcomp}},
 \code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp.noboot}},
 but is documented here for clarity. Generally, users will not need to call
 this function directly.
}
