% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_multinomial.R
\name{homogeneity_test.qgcompmultfit}
\alias{homogeneity_test.qgcompmultfit}
\title{Hypothesis testing about joint effect of exposures on a multinomial outcome}
\usage{
\method{homogeneity_test}{qgcompmultfit}(x, ...)
}
\arguments{
\item{x}{Result from qgcomp multinomial fit (qgcompmultfit object).}

\item{...}{Unused}
}
\value{
qgcompmulttest object (list) with results of a chi-squared test
}
\description{
Tests the null hypothesis that the joint effect of the mixture components is identical across all referent outcome types (homogeneity test for linear effect of the mixture on a quantized basis)
}
