% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pointwise_total.boot}
\alias{pointwise_total.boot}
\title{pointwise_total: calculating pointwise comparisons for qgcomp.boot objects}
\usage{
pointwise_total.boot(x, pointwiseref = 1)
}
\arguments{
\item{x}{"qgcompfit" object from `qgcomp.boot`,}

\item{pointwiseref}{referent quantile (e.g. 1 uses the lowest joint-exposure category as 
the referent category for calculating all mean differences/standard deviations)}
}
\description{
Calculates: expected outcome (on the link scale), mean difference (link scale)
and the standard deviation of the mean difference (link scale) for pointwise comparisons
}
\examples{
set.seed(12)
\donttest{
dat <- data.frame(x1=(x1 <- runif(50)), x2=runif(50), x3=runif(50), z=runif(50),
                  y=runif(50)+x1+x1^2)
ft <- qgcomp.boot(y ~ z + x1 + x2 + x3, expnms=c('x1','x2','x3'), data=dat, q=10)
pointwise_total.boot(ft, 4)
}
}
\seealso{
\code{\link[qgcomp]{qgcomp.boot}}
}
