% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_experimental.R
\name{print.qgcomppartialavg}
\alias{print.qgcomppartialavg}
\title{Default printing method for a qgcomppartialavg object}
\usage{
\method{print}{qgcomppartialavg}(x, ...)
}
\arguments{
\item{x}{"qgcompfit" object from \code{.qgcomp.partials_avg}
#' function}
}
\description{
Default printing method for a qgcomppartialavg object
}
