% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgcomp_long.R
\name{qgcomp.gee.boot}
\alias{qgcomp.gee.boot}
\title{estimation of quantile g-computation fit, using bootstrap confidence intervals}
\usage{
qgcomp.gee.boot(f, data, expnms = NULL, q = 4, breaks = NULL,
  corstr = "independence", id = NULL, alpha = 0.05, B = 200,
  rr = TRUE, degree = 1, seed = NULL, ...)
}
\arguments{
\item{f}{R style formula}

\item{data}{data frame}

\item{expnms}{character vector of exposures of interest}

\item{q}{NULL or number of quantiles used to create quantile indicator variables
representing the exposure variables. If NULL, then gcomp proceeds with un-transformed
version of exposures in the input datasets (useful if data are already transformed,
or for performing standard g-computation)}

\item{breaks}{(optional) NULL, or a list of (equal length) numeric vectors that 
characterize the minimum value of each category for which to 
break up the variables named in expnms. This is an alternative to using 'q'
to define cutpoints.}

\item{corstr}{character string with the type of correlation structure implied
by the working covariance matrix: "independence", "exchangeable", "ar1", 
"unstructured" and "userdefined" (See \code{\link[geepack]{geeglm}})}

\item{id}{(optional) NULL, or variable name indexing individual units of 
observation (only needed if analyzing data with multiple observations per 
id/cluster)}

\item{alpha}{alpha level for confidence limit calculation}

\item{B}{integer: number of bootstrap iterations (this should typically be
>=200, though it is set lower in examples to improve run-time).}

\item{rr}{logical: if using binary outcome and rr=TRUE, qgcomp.boot will 
estimate risk ratio rather than odds ratio}

\item{degree}{polynomial basis function for marginal model (e.g. degree = 2
allows that the relationship between the whole exposure mixture and the outcome
is quadratic.}

\item{seed}{integer or NULL: random number seed for replicable bootstrap results}

\item{...}{arguments to glm (e.g. family)}
}
\value{
a qgcompfit object, which contains information about the effect
 measure of interest (psi) and associated variance (var.psi), as well
 as information on the model fit (fit) and information on the 
 marginal structural model (msmfit) used to estimate the final effect
 estimates.
}
\description{
This function yields population average effect estimates for 
  both continuous and binary outcomes
}
\details{
Estimates correspond to the average expected change in the
 (log) outcome per quantile increase in the joint exposure to all exposures 
 in `expnms'. Test statistics and confidence intervals are based on 
 a non-parametric bootstrap, using the standard deviation of the bootstrap
 estimates to estimate the standard error. The bootstrap standard error is 
 then used to estimate Wald-type confidence intervals. Note that no bootstrapping
 is done on estimated quantiles of exposure, so these are treated as fixed
 quantities
}
\examples{
set.seed(30)
geedat <- data.frame(i = round((1:50)/4), x1=runif(50), x2=runif(50), z=runif(50))
geedat <- within(geedat, {y=rbinom(50, 1, 1/(1+exp(-log((i+1)/10))))})
qgcomp.boot(y ~ z + x1 + x2, family="binomial", expnms = c('x1', 'x2'), 
  data=geedat, q=2, B=10, id='i')
mod = qgcomp.gee.boot(f = y ~ z + x1 + x2, family="binomial", expnms = c('x1', 'x2'), 
  data=geedat, q=2, B=100, id='i', corstr="independence")
mod2 = qgcomp.gee.boot(y ~ z + x1 + x2 + I(x1*x2), family="binomial", expnms = c('x1', 'x2'), 
  data=geedat, q=2, B=100, id='i', corstr="independence", rr=FALSE)
# compared with standard qgcomp
set.seed(30)
geedat <- data.frame(i = round((1:100)/4), x1=runif(100), x2=runif(100), z=runif(100))
geedat <- within(geedat, {y=rbinom(100, 1, 1/(1+exp(-log((i+1)/10))))})
mod3 <- qgcomp.gee.boot(y ~ z + x1 + x2+ I(x1*x2), family=binomial(), expnms = c('x1', 'x2'), 
  data=geedat, q=4, B=10, rr=FALSE, degree=2, id='i', corstr="exchangeable")
plot(mod3)
mod4 <- qgcomp.boot(y ~ z + x1 + x2+ I(x1*x2), family=binomial(), expnms = c('x1', 'x2'), 
  data=geedat, q=4, B=10, rr=FALSE, degree=2, id='i')
plot(mod4)
}
\seealso{
\code{\link[qgcomp]{qgcomp.noboot}}, and \code{\link[qgcomp]{qgcomp}}
}
\keyword{mixtures}
\keyword{variance,}
