% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qgcomp.noboot}
\alias{qgcomp.noboot}
\title{qgcomp::qgcomp.noboot}
\usage{
qgcomp.noboot(f, data, expcoefs = NULL, q = 4, alpha = 0.05, ...)
}
\arguments{
\item{f}{R style formula}

\item{data}{data frame}

\item{expcoefs}{logical/numeric vector that points to variables in the
formula that should be considered part of the 'mixture'}

\item{q}{number of quantiles used to create quantile indicator variables
representing the exposure variables}

\item{alpha}{alpha level for confidence limit calculation}

\item{...}{arguments to glm (e.g. family)}
}
\description{
create variables representing indicator functions with cutpoints defined
by quantiles
}
\examples{
dat = data.frame(y=runif(10), x1=runif(10), x2=runif(10), z=runif(10))
qgcomp.noboot(y ~ z + x1 + x2, expcoefs=c(0,1,1), data=dat, q=2)
}
\keyword{mixtures}
\keyword{variance,}
