% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qgcomp}
\alias{qgcomp}
\title{qgcomp::qgcomp: estimation of quantile g-computation fit}
\usage{
qgcomp(f, data = data, family = gaussian(), rr = FALSE, ...)
}
\arguments{
\item{f}{R style formula}

\item{data}{data frame}

\item{family}{`gaussian()` or `binomial()`}

\item{rr}{logical: if using binary outcome and rr=TRUE, qgcomp.boot will estimate risk ratio rather than odds ratio}

\item{...}{arguments to qgcomp.noboot or qgcomp.boot (e.g. q)}
}
\description{
qgcomp::qgcomp: estimation of quantile g-computation fit
}
\examples{
set.seed(50)
dat = data.frame(y=runif(50), x1=runif(50), x2=runif(50), z=runif(50))
qgcomp.noboot(y ~ z + x1 + x2, expnms = c('x1', 'x2'), data=dat, q=2)
qgcomp.boot(y ~ z + x1 + x2, expnms = c('x1', 'x2'), data=dat, q=2, B=100)
}
\keyword{mixtures}
\keyword{variance,}
