% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_multinomial.R
\name{summary.qgcompmultfit}
\alias{summary.qgcompmultfit}
\title{Summarize gcompmultfit object}
\usage{
\method{summary}{qgcompmultfit}(object, ..., tests = NULL)
}
\arguments{
\item{object}{Result from qgcomp multinomial fit (qgcompmultfit object).}

\item{...}{Unused}

\item{tests}{Character vector (e.g. c("global", "homogeneity")) that determine the types of hypothesis tests that are printed}
}
\value{
qgcompmulttest object (list) with results of a chi-squared test
}
\description{
Summary printing to include coefficients, standard errors, hypothesis tests, weights
}
