% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_extensions.R
\name{glance.qgcompfit}
\alias{glance.qgcompfit}
\title{Glance at a qgcompfit object}
\usage{
\method{glance}{qgcompfit}(x, ...)
}
\arguments{
\item{x}{a qgcompfit object}

\item{...}{Not used}
}
\description{
Glance accepts a model object and returns a
tibble::tibble() with exactly one row of model summaries. The summaries are
typically goodness of fit measures, p-values for hypothesis tests on residuals,
or model convergence information.

Glance never returns information from the original call to the modeling function.
This includes the name of the modeling function or any arguments passed to the
modeling function.

Glance does not calculate summary measures. Rather, it farms out these computations
to appropriate methods and gathers the results together. Sometimes a goodness of
fit measure will be undefined. In these cases the measure will be reported as NA.
(Description taken from \code{broom::glance} help file.)
}
