% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_multinomial.R
\name{homogeneity_test}
\alias{homogeneity_test}
\title{Hypothesis testing about joint effect of exposures on a multinomial outcome}
\usage{
homogeneity_test(x, ...)
}
\arguments{
\item{x}{Result from fit.}

\item{...}{Unused}
}
\description{
Hypothesis testing about joint effect of exposures on a multinomial outcome
}
