% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msm.fit}
\alias{msm.fit}
\title{fitting marginal structural model (MSM) based on g-computation with
quantized exposures}
\usage{
msm.fit(f, qdata, q, expnms, rr = TRUE, main = TRUE, ...)
}
\arguments{
\item{f}{an r function representing the conditional model for the outcome, given all
exposures and covariates}

\item{q}{integer, number of quantiles used in creating quantized variables}

\item{expnms}{a character vector with the names of  the columns to be
quantized}

\item{rr}{logical, estimate log(risk ratio) (family='binomial' only)}

\item{main}{logical, internal use: produce estimates of exposure effect (gamma)
and expected outcomes under g-computation and the MSM}

\item{data}{a data frame}
}
\description{
this is an internal function called by \code{\link[qgcomp]{qgcomp}},
 \code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp.noboot}},
 but is documented here for clarity. Generally, users will not need to call
 this function directly.
}
\details{
This function first computes expected outcomes under hypothetical
interventions to simultaneously set all exposures to a specific quantile. These
predictions are based on g-computation, where the exposures are `quantized',
meaning that they take on ordered integer values according to their ranks,
and the integer values are determined by the number of quantile cutpoints used.
The function then takes these expected outcomes and fits an additional model
(a marginal structural model) with the expected outcomes as the outcome and
the intervention value of the exposures (the quantile integer) as the exposure.
Under causal identification assumptions and correct model specification,
the MSM yields a causal exposure-response representing the incremental
change in the expected outcome given a joint intervention on all exposures.
}
\seealso{
\code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp}}
}
\keyword{mixtures}
\keyword{variance,}
