% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{msm.fit}
\alias{msm.fit}
\title{fitting marginal structural model (MSM) based on g-computation with
quantized exposures}
\usage{
msm.fit(f, qdata, intvals, expnms, rr = TRUE, main = TRUE,
  degree = 1, id = NULL, bayes, ...)
}
\arguments{
\item{f}{an r formula representing the conditional model for the outcome, given all
exposures and covariates. Interaction terms that include exposure variables
should be represented via the \code{\link[base]{I}} function}

\item{qdata}{a data frame with quantized exposures}

\item{intvals}{sequence, the sequence of integer values that the joint exposure 
is 'set' to for estimating the msm. For quantile g-computation, this is just 
0:(q-1), where q is the number of quantiles of exposure.}

\item{expnms}{a character vector with the names of the columns in qdata that represent
the exposures of interest (main terms only!)}

\item{rr}{logical, estimate log(risk ratio) (family='binomial' only)}

\item{main}{logical, internal use: produce estimates of exposure effect (psi)
and expected outcomes under g-computation and the MSM}

\item{degree}{polynomial basis function for marginal model (e.g. degree = 2
allows that the relationship between the whole exposure mixture and the outcome
is quadratic. Default=1)}

\item{id}{(optional) NULL, or variable name indexing individual units of 
observation (only needed if analyzing data with multiple observations per 
id/cluster)}

\item{bayes}{use underlying Bayesian model (`arm` package defaults). Results
in penalized parameter estimation that can help with very highly correlated 
exposures. Note: this does not lead to fully Bayesian inference in general, 
so results should be interpereted as frequentist.}

\item{...}{arguments to glm (e.g. family)}
}
\description{
This is an internal function called by \code{\link[qgcomp]{qgcomp}},
 \code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp.noboot}},
 but is documented here for clarity. Generally, users will not need to call
 this function directly.
}
\details{
This function first computes expected outcomes under hypothetical
interventions to simultaneously set all exposures to a specific quantile. These
predictions are based on g-computation, where the exposures are `quantized',
meaning that they take on ordered integer values according to their ranks,
and the integer values are determined by the number of quantile cutpoints used.
The function then takes these expected outcomes and fits an additional model
(a marginal structural model) with the expected outcomes as the outcome and
the intervention value of the exposures (the quantile integer) as the exposure.
Under causal identification assumptions and correct model specification,
the MSM yields a causal exposure-response representing the incremental
change in the expected outcome given a joint intervention on all exposures.
}
\examples{
set.seed(50)
dat <- data.frame(y=runif(200), x1=runif(200), x2=runif(200), z=runif(200))
X <- c('x1', 'x2')
qdat <- quantize(dat, X, q=4)$data
mod <- qgcomp:::msm.fit(f=y ~ z + x1 + x2 + I(x1*x2), 
        expnms = c('x1', 'x2'), qdata=qdat, intvals=1:4, bayes=FALSE)
summary(mod$fit) # outcome regression model
summary(mod$msmfit) # msm fit (variance not valid - must be obtained via bootstrap)
}
\seealso{
\code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp}}
}
\concept{variance mixtures}
