% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msm.predict}
\alias{msm.predict}
\title{plot.qgcompfit: default prediction method for a qgcompfit object}
\usage{
msm.predict(object, newdata = NULL)
}
\arguments{
\item{object}{"qgcompfit" object from `qgcomp.boot` function}

\item{newdata}{(optional) new set of data (data frame) with a variable 
called `gamma` representing the joint exposure level of all exposures
under consideration
set.seed(50)
dat <- data.frame(y=runif(50), x1=runif(50), x2=runif(50), z=runif(50))
obj <- qgcomp.boot(y ~ z + x1 + x2 + I(z*x1), expnms = c('x1', 'x2'), data=dat, q=4, B=100)
dat2 <- data.frame(gamma=seq(1,4, by=0.1))
summary(msm.predict(obj))
summary(msm.predict(obj, newdata=dat2))}

\item{expnms}{character vector of exposures of interest}
}
\description{
get predicted values from a qgcompfit object from
\code{\link[qgcomp]{qgcomp.boot}}
}
