% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.qgcompfit}
\alias{plot.qgcompfit}
\title{plot.qgcompfit: default plotting method for a qgcompfit object}
\usage{
\method{plot}{qgcompfit}(x, ...)
}
\arguments{
\item{x}{"qgcompfit" object from `qgcomp.noboot` or  `qgcomp.boot` functions}

\item{...}{unused}
}
\description{
Plot a quantile g-computation object. For qgcomp.noboot, this function will
createa butterfly plot of weights. For qgcomp.boot, this function will create
a box plot with smoothed line overlaying that represents a non-parametric
fit of a model to the expected outcomes in the population at each quantile
of the joint exposures (e.g. '1' represents 'at the first quantile for
every exposure')
}
\examples{
set.seed(12)
dat <- data.frame(x1=(x1 <- runif(100)), x2=runif(100), x3=runif(100), z=runif(100),y=runif(100)+x1^2)
ft <- qgcomp.noboot(y ~ z + x1 + x2 + x3, expnms=c('x1','x2','x3'), data=dat, q=4)
plot(ft)

# using non-linear outcome model
ft2 <- qgcomp.boot(y ~ z + x1 + x2 + x3 + I(x1*x1), expnms=c('x1','x2','x3'), 
data=dat, q=4, B=10)
plot(ft2)

# using non-linear marginal structural model
ft2 <- qgcomp.boot(y ~ z + x1 + x2 + x3 + I(x1*x1), expnms=c('x1','x2','x3'), 
data=dat, q=4, B=10, degree=2)
plot(ft2)
}
\seealso{
\code{\link[qgcomp]{qgcomp.noboot}}, \code{\link[qgcomp]{qgcomp.boot}}, and \code{\link[qgcomp]{qgcomp}}
}
