% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.qgcompfit}
\alias{plot.qgcompfit}
\title{qgcomp::plot.qgcompfit}
\usage{
\method{plot}{qgcompfit}(x, ...)
}
\arguments{
\item{x}{"qgcompfit" object from `qgcomp.noboot` or  `qgcomp.boot` functions}

\item{...}{unused}
}
\description{
plot quantile g-computation object
}
\examples{
set.seed(12)
dat = data.frame(y=runif(100), x1=runif(100), x2=runif(100), z=runif(100))
ft = qgcomp.noboot(y ~ z + x1 + x2, expnms=c('x1','x2'), data=dat, q=2)
plot(ft)
ft2 = qgcomp.boot(y ~ z + x1 + x2, expnms=c('x1','x2'), data=dat, q=2)
plot(ft2)
}
\keyword{mixtures}
\keyword{variance,}
