% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.qgcompfit}
\alias{print.qgcompfit}
\title{qgcomp::print.qgcompfit}
\usage{
\method{print}{qgcompfit}(x, ...)
}
\arguments{
\item{x}{"qgcompfit" object from `qgcomp.noboot` function}

\item{...}{unused}
}
\description{
carry out quantile g-computation
by quantiles
}
\examples{
dat = data.frame(y=runif(10), x1=runif(10), x2=runif(10), z=runif(10))
qgcomp.noboot(y ~ z + x1 + x2, expcoefs=c(0,1,1), data=dat, q=2)
}
\keyword{mixtures}
\keyword{variance,}
