% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_experimental.R
\name{print.qgcomppartialavg_boot}
\alias{print.qgcomppartialavg_boot}
\title{Default printing method for a qgcomppartialavg object}
\usage{
\method{print}{qgcomppartialavg_boot}(x, ...)
}
\description{
Default printing method for a qgcomppartialavg object
}
