% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qgcomp.noboot}
\alias{qgcomp.noboot}
\title{qgcomp::qgcomp.noboot}
\usage{
qgcomp.noboot(f, data, expnms = NULL, q = 4, alpha = 0.05, ...)
}
\arguments{
\item{f}{R style formula}

\item{data}{data frame}

\item{expnms}{character vector of exposures of interest}

\item{q}{number of quantiles used to create quantile indicator variables
representing the exposure variables}

\item{alpha}{alpha level for confidence limit calculation}

\item{...}{arguments to glm (e.g. family)}
}
\description{
create variables representing indicator functions with cutpoints defined
by quantiles
}
\examples{
set.seed(50)
dat = data.frame(y=runif(50), x1=runif(50), x2=runif(50), z=runif(50))
qgcomp.noboot(f=y ~ z + x1 + x2, expnms = c('x1', 'x2'), data=dat, q=2)
}
\keyword{mixtures}
\keyword{variance,}
