% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quantize}
\alias{quantize}
\title{create variables representing indicator functions with cutpoints defined
by quantiles}
\usage{
quantize(data, expnms, q)
}
\arguments{
\item{data}{a data frame}

\item{expnms}{a character vector with the names of  the columns to be
quantized}

\item{q}{integer, number of quantiles used in creating quantized variables}
}
\description{
create variables representing indicator functions with cutpoints defined
by quantiles
}
\examples{
dat = data.frame(y=runif(10), x1=runif(10), x2=runif(10), z=runif(10))
qdata = quantize(data=dat, expnms=c("x1", "x2"), q=2)
}
\keyword{mixtures}
\keyword{variance,}
