% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{se_comb}
\alias{se_comb}
\title{qgcomp::se_comb}
\usage{
se_comb(weightvec, covmat)
}
\arguments{
\item{weightvec}{a logical index of which columns are used to create
composite effect estimate}

\item{covmat}{covariance matrix from a glm fit}
}
\description{
calculate standard error of weighted linear combination of random variables
 given a vector of weights and a covariance matrix
}
\examples{
dat = data.frame(y=runif(10), x=runif(10))
lmfit = glm(y ~ x, data=dat, family='gaussian')
qgcomp:::se_comb(weightvec=c(0,1), covmat=summary(lmfit)$cov.scaled)
}
\keyword{mixtures}
\keyword{variance,}
